fixed the usage of UI modal in filter list

+++ Modules/Bibliographic/classes/FieldFilter/class.ilBiblFieldFilterGUI.php
@@ -74,16 +74,8 @@ class ilBiblFieldFilterGUI {
 
 	public function index() {
 		if ($this->access()->checkAccess('write', "", $this->facade->iliasRefId())) {
-			$ilBiblSettingsFilterFormGUI = new ilBiblFieldFilterFormGUI($this, new ilBiblFieldFilter(), $this->facade);
-
-			$f = $this->dic()->ui()->factory();
-			$r = $this->dic()->ui()->renderer();
-			$modal = $f->modal()->roundtrip($this->lng()->txt("add_filter"), $f->legacy($ilBiblSettingsFilterFormGUI->getHTML()));
-			$button = $f->button()->standard($this->lng()->txt("add_filter"), "#")->withOnClick($modal->getShowSignal());
-
-			$add_filter_html = $r->render([$modal, $button]);
-
-			$this->toolbar()->addText($add_filter_html);
+			$button = $this->dic()->ui()->factory()->button()->primary($this->lng()->txt("add_filter"), $this->ctrl()->getLinkTarget($this, self::CMD_ADD));
+			$this->toolbar()->addText($this->dic()->ui()->renderer()->render([$button]));
 		}
 
 		$table = new ilBiblFieldFilterTableGUI($this, $this->facade);
@@ -117,13 +109,6 @@ class ilBiblFieldFilterGUI {
 	}
 
 
-	public function renderEditFormAsync() {
-		$ilBiblSettingsFilterFormGUI = $this->initEditForm();
-		echo $ilBiblSettingsFilterFormGUI->getHTML();
-		exit;
-	}
-
-
 	public function update() {
 		$il_bibl_field = $this->getFieldFilterFromRequest();
 		$this->tabs()->activateTab(self::CMD_STANDARD);
@@ -140,9 +125,14 @@ class ilBiblFieldFilterGUI {
 
 	public function delete() {
 		global $DIC;
-		$il_bibl_field = $this->getFieldFilterFromRequest();
-		$this->tabs()->activateTab(self::CMD_STANDARD);
-		$il_bibl_field->delete();
+		//		$this->tpl()->setContent('<pre>' . print_r($_POST, 1) . '</pre>');
+		$items = $this->http()->request()->getParsedBody()['interruptive_items'];
+		if (is_array($items)) {
+			foreach ($items as $filter_id) {
+				$il_bibl_field = $this->facade->filterFactory()->findById($filter_id);
+				$il_bibl_field->delete();
+			}
+		}
 		ilUtil::sendSuccess($DIC->language()->txt('filter_deleted'), true);
 		$this->ctrl()->redirect($this, self::CMD_STANDARD);
 	}
